Early_alzheimer_detection_backend/Early_alzheimer_detection_b - Copy/requirements.txt
asgiref==3.6.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
fonttools==4.39.4
kiwisolver==1.4.4
matplotlib==3.7.1
nibabel==5.1.0
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pydicom==2.3.1
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
vtk==9.2.6
